Move to_one_line_messages to sourceline.py
