+++ schema_salad/main.py
@@ -2,12 +2,10 @@
 from __future__ import absolute_import, print_function
 
 import argparse
-import itertools
 import logging
 import os
-import re
 import sys
-from typing import (Any, Dict, List, Mapping, MutableSequence, Pattern, Tuple,
+from typing import (Any, Dict, List, Mapping, MutableSequence,
                     Union, cast)
 
 import pkg_resources  # part of setuptools
@@ -21,7 +19,7 @@ from typing_extensions import Text  # pylint: disable=unused-import
 
 from . import codegen, jsonld_context, schema, validate
 from .ref_resolver import Loader, file_uri
-from .sourceline import strip_dup_lineno
+from .sourceline import strip_dup_lineno, to_one_line_messages, reformat_yaml_exception_message
 from .utils import json_dumps
 from .avro.schema import SchemaParseException
 
@@ -37,84 +35,6 @@ def printrdf(workflow,  # type: str
     # type: (...) -> None
     g = jsonld_context.makerdf(workflow, wf, ctx)
     print(g.serialize(format=sr, encoding='utf-8').decode('utf-8'))  # type: ignore
-
-def regex_chunk(lines, regex):
-    # type: (List[str], Pattern[str]) -> List[List[str]]
-    lst = list(itertools.dropwhile(lambda x: not regex.match(x), lines))
-    arr = []
-    while lst:
-        ret = [lst[0]]+list(itertools.takewhile(lambda x: not regex.match(x),
-                                                lst[1:]))
-        arr.append(ret)
-        lst = list(itertools.dropwhile(lambda x: not regex.match(x),
-                                       lst[1:]))
-    return arr
-
-
-def chunk_messages(message):  # type: (str) -> List[Tuple[int, str]]
-    file_regex = re.compile(r'^(.+:\d+:\d+:)(\s+)(.+)$')
-    item_regex = re.compile(r'^\s*\*\s+')
-    arr = []
-    for chun in regex_chunk(message.splitlines(), file_regex):
-        fst = chun[0]
-        mat = file_regex.match(fst)
-        if mat:
-            place = mat.group(1)
-            indent = len(mat.group(2))
-
-            lst = [mat.group(3)]+chun[1:]
-            if [x for x in lst if item_regex.match(x)]:
-                for item in regex_chunk(lst, item_regex):
-                    msg = re.sub(item_regex, '', "\n".join(item))
-                    arr.append((indent, place+' '+re.sub(
-                        r'[\n\s]+', ' ', msg)))
-            else:
-                msg = re.sub(item_regex, '', "\n".join(lst))
-                arr.append((indent, place+' '+re.sub(
-                    r'[\n\s]+', ' ', msg)))
-    return arr
-
-
-def to_one_line_messages(message):  # type: (str) -> str
-    ret = []
-    max_elem = (0, '')
-    for (indent, msg) in chunk_messages(message):
-        if indent > max_elem[0]:
-            max_elem = (indent, msg)
-        else:
-            ret.append(max_elem[1])
-            max_elem = (indent, msg)
-    ret.append(max_elem[1])
-    return "\n".join(ret)
-
-
-def reformat_yaml_exception_message(message):  # type: (str) -> str
-    line_regex = re.compile(r'^\s+in "(.+)", line (\d+), column (\d+)$')
-    fname_regex = re.compile(r'^file://'+re.escape(os.getcwd())+'/')
-    msgs = message.splitlines()
-    ret = []
-
-    if len(msgs) == 3:
-        msgs = msgs[1:]
-        nblanks = 0
-    elif len(msgs) == 4:
-        c_msg = msgs[0]
-        match = line_regex.match(msgs[1])
-        if match:
-            c_file, c_line, c_column = match.groups()
-            c_file = re.sub(fname_regex, '', c_file)
-            ret.append("%s:%s:%s: %s" % (c_file, c_line, c_column, c_msg))
-
-        msgs = msgs[2:]
-        nblanks = 2
-
-    p_msg = msgs[0]
-    match = line_regex.match(msgs[1])
-    if match:
-        p_file, p_line, p_column = match.groups()
-        p_file = re.sub(fname_regex, '', p_file)
-        ret.append("%s:%s:%s:%s %s" % (p_file, p_line, p_column, ' '*nblanks, p_msg))
-    return "\n".join(ret)
 
 
 def main(argsl=None):  # type: (List[str]) -> int
